requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.7
djangorestframework==3.15.2
django_elasticsearch_dsl
elastic-transport==7.15.0
elasticsearch==7.15.0
elasticsearch-dsl==7.4.1
flower==2.0.1
humanize==4.10.0
kombu==5.3.7
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
vine==5.1.0
wcwidth==0.2.13
